KST-GCN/requirements.txt
tensorflow-gpu==1.15.0
pandas==1.1.5
numpy==1.21.5
matplotlib==3.5.3
scipy==1.7.3
scikit-learn==1.0.2
protobuf==3.20.3